+++ board/main.c
@@ -12,7 +12,7 @@
 #include "drivers/registers.h"
 #include "drivers/interrupts.h"
 
-#include "drivers/llcan.h"
+#include "stm32fx/llcan.h"
 #include "drivers/llgpio.h"
 #include "stm32fx/lladc.h"
 #include "drivers/pwm.h"
