+++ board/spi_flasher.h
@@ -153,7 +153,7 @@ int spi_cb_rx(uint8_t *data, int len, uint8_t *data_out) {
 
 #ifdef PEDAL
 
-#include "drivers/llcan.h"
+#include "stm32fx/llcan.h"
 #define CAN CAN1
 
 #define CAN_BL_INPUT 0x1
